+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# UNRELEASED
+
+## Changes
+- Code structure changes
+  - all routes are now in their own separate static handler, woo! HTTPApi is no longer a 600+ line long file.
+
 # `0.6.6`
 
 ## Changes